// React/react-test-project/src/shared/API/Enums.ts
export enum ButtonVariant {
    TEXT = 'text',
    OUTLINED = 'outlined',
    BASIC = 'basic'
}

export enum ButtonColor {
    PRIMARY = '#6366F1',
    DANGER = '#EF4444',
    SECONDARY = '#64748B',
    SUCCESS = '#22C55E',
    WARNING = '#F59E0B'
}

export enum ButtonSize {
    SMALL = 'small',
    MEDIUM = 'medium',
    LARGE = 'large'
}